+++ src/componentes/Globais/SidebarLeft/index.js
@@ -107,7 +107,7 @@ export const SidebarLeft = () => {
                         </NavItem>
                         )
                     }
-                    
+
 
                     {/*
